requirements.txt
# Core Dependencies for YOLO Ingredient Scanner

# Machine Learning & Computer Vision
ultralytics>=8.0.0           # YOLO model framework
opencv-python>=4.8.0         # Computer vision and camera operations
numpy>=1.21.0                # Numerical computations
torch>=1.11.0                # PyTorch (required by ultralytics)
torchvision>=0.12.0          # PyTorch vision utilities
Pillow>=9.0.0                # Image processing

# Web Framework
Flask>=2.3.0                 # Web framework for the application

# System & Utilities
psutil>=5.9.0                # System monitoring (optional but useful)

# Optional Performance Enhancements
# Uncomment these for better performance:
# opencv-contrib-python>=4.8.0  # Additional OpenCV features
# onnxruntime>=1.12.0            # ONNX runtime for optimized inference

# Development & Testing (optional)
# pytest>=7.0.0                 # For running tests
# requests>=2.28.0              # For API testing